+++ R/zzz.R
@@ -11,5 +11,5 @@ if(getRversion() >= "2.15.1")
       "archigos", "atop_alliance", "gw_sdp_gdp", "cow_sdp_gdp", "ucdp_onsets",
       "ucdp_acd", "cow_mid_dirdisps", "cow_mid_disps", "cow_mid_ddydisps",
       "gml_mid_ddydisps", "td_rivalries", "gw_cow_years", "rugged", "creg",
-      "hief", "gwcode_democracy", "cow_war_intra")
+      "hief", "gwcode_democracy", "cow_war_intra", "cow_war_inter")
   )