+++ data-raw/cow_wars/cow_intrastate_wars.R
@@ -27,7 +27,7 @@ cow_war_intra_raw %>%
   select(warnum:warname, year, wartype:intnl, outcome, sideadeaths, sidebdeaths, ongo2007) %>%
   arrange(warnum, year) %>%
   group_by(warnum) %>%
-  mutate(cowcwonset = ifelse(row_number() == 1, 1, 0)) %>% ungroup() -> cow_war_intra
+  mutate(cowintraonset = ifelse(row_number() == 1, 1, 0)) %>% ungroup() -> cow_war_intra
 
 
 
@@ -38,14 +38,60 @@ combat_resumes %>%
   select(warnum:warname, year, wartype:intnl, outcome, sideadeaths, sidebdeaths, ongo2007, resume_combat) %>%
   arrange(warnum, year) %>%
   group_by(warnum) %>%
-  mutate(cowcwonset = ifelse(row_number() == 1, 1, 0))  %>% ungroup() %>%
+  mutate(cowintraonset = ifelse(row_number() == 1, 1, 0))  %>% ungroup() %>%
   bind_rows(cow_war_intra, .) %>%
   mutate(resume_combat = ifelse(is.na(resume_combat), 0, resume_combat)) %>%
   arrange(warnum, -primary_state, year) -> cow_war_intra
 
 
 cow_war_intra %>%
-  mutate(cowcwongoing = 1) %>%
-  select(warnum, warname, wartype, year, cowcwonset, cowcwongoing, resume_combat, everything()) -> cow_war_intra
+  mutate(cowintraongoing = 1) %>%
+  select(warnum, warname, wartype, year, cowintraonset, cowintraongoing, resume_combat, everything()) -> cow_war_intra
 
 save(cow_war_intra, file="data/cow_war_intra.rda")
+
+
+# fiddle with things...
+
+cow_war_intra %>%
+  rename(ccode = ccodea) %>%
+  filter(primary_state == 1) %>%
+  # filter(wartype == "central control") %>%
+  group_by(ccode, year) %>%
+  mutate(intrawarnums = paste0(.data$warnum, collapse = "; ")) %>%
+  #filter(n() > 1) %>%
+  select(warnum:wartype, ccode, year, everything()) %>%
+  mutate(duplicated = ifelse(n() > 1, 1, 0),
+         sd = sd(cowcwonset)) %>%
+  # Remove anything that's not a unique war onset
+  mutate(removeme = case_when(
+    duplicated == 1 & sd > 0 & cowcwonset == 0 ~ 1,
+    TRUE ~ 0
+  )) %>%
+  filter(removeme == 0) %>%
+  group_by(ccode, year) %>%
+  # filter(n() > 1) %>%
+  mutate_at(vars(contains("deaths")), ~ifelse(. < 0, NA, .)) %>%
+  rowwise() %>%
+  mutate(deaths = sideadeaths + sidebdeaths) %>%
+  mutate(deaths = case_when(
+    is.na(deaths) & is.na(sidebdeaths) ~ sideadeaths,
+    is.na(deaths) & is.na(sideadeaths) ~ sidebdeaths,
+    TRUE ~ deaths
+  )) %>%
+  group_by(ccode, year) %>%
+  mutate(duplicated = ifelse(n() > 1, 1, 0)) %>%
+  arrange(-deaths) %>%
+  slice(1) %>%
+  ungroup() %>%
+  select(warnum:cowcwongoing, intnl, outcome, sideadeaths, sidebdeaths) %>%
+  # selecting on primary states means there are no -8s, just -9s where missing
+  mutate_at(vars(contains("deaths")), ~ifelse(is.na(.), -9, .)) -> the_data
+
+
+
+cow_war_intra %>%
+  rename(ccode = ccodea) %>%
+  filter(primary_state == 1) %>%
+  filter(wartype == "central control") %>%
+  mutate_at(vars(contains("deaths")), ~ifelse(. < 0, NA, .)) %>% data.frame